@@ -1,5 +1,5 @@
 {
-    "name": "buckhamduffy/scout-elasticsearch-driver",
+    "name": "theporndb/scout-elasticsearch-driver",
     "type": "library",
     "description": "The Elasticsearch Driver for Laravel Scout",
     "keywords": [
@@ -26,16 +26,17 @@
         }
     ],
     "require": {
-        "php": "^8.0",
-        "elasticsearch/elasticsearch": ">=7.0 <8",
-        "laravel/scout": "^9.0|^10.0",
-        "illuminate/contracts": "^9.0|^10.0|^11.0",
-        "ext-json": "*"
+        "php": "^8.4",
+        "ext-json": "*",
+        "elasticsearch/elasticsearch": "^8.0",
+        "illuminate/contracts": "^11.0",
+        "laravel/scout": "^10.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0|^8.0",
+        "laravel/pint": "^v1.16",
         "mockery/mockery": "^1.0",
-        "buckhamduffy/coding-standards": "^3.2"
+        "phpstan/phpstan": "^2.0",
+        "phpunit/phpunit": "^11.0"
     },
     "autoload": {
         "psr-4": {
@@ -48,6 +49,11 @@
             "providers": [
                 "ScoutElastic\\ScoutElasticServiceProvider"
             ]
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "php-http/discovery": true
         }
     }
 }
