@@ -1,12 +1,12 @@
 <?php
 
 return [
-	'client' => [
-		'hosts' => [
-			env('SCOUT_ELASTIC_HOST', 'localhost:9200'),
-		],
-	],
-	'update_mapping'   => env('SCOUT_ELASTIC_UPDATE_MAPPING', true),
-	'indexer'          => env('SCOUT_ELASTIC_INDEXER', 'single'),
-	'document_refresh' => env('SCOUT_ELASTIC_DOCUMENT_REFRESH')
+    'client' => [
+        'hosts' => [
+            env('SCOUT_ELASTIC_HOST', 'localhost:9200'),
+        ],
+    ],
+    'update_mapping'   => env('SCOUT_ELASTIC_UPDATE_MAPPING', true),
+    'indexer'          => env('SCOUT_ELASTIC_INDEXER', 'single'),
+    'document_refresh' => env('SCOUT_ELASTIC_DOCUMENT_REFRESH')
 ];
