@@ -8,39 +8,39 @@ use ScoutElastic\Console\Features\RequiresModelArgument;
 
 class ElasticCompareModelCommand extends Command
 {
-	use RequiresModelArgument {
-		RequiresModelArgument::getArguments as private modelArgument;
-	}
+    use RequiresModelArgument {
+        RequiresModelArgument::getArguments as private modelArgument;
+    }
 
-	/**
-	 * {@inheritDoc}
-	 * @var string
-	 */
-	protected $name = 'elastic:compare-model';
+    /**
+     * {@inheritDoc}
+     * @var string
+     */
+    protected $description = 'Compare the count of the model with the count of the index';
 
-	/**
-	 * {@inheritDoc}
-	 * @var string
-	 */
-	protected $description = 'Compare the count of the model with the count of the index';
+    /**
+     * {@inheritDoc}
+     * @var string
+     */
+    protected $name = 'elastic:compare-model';
 
-	/**
-	 * Handle the command.
-	 */
-	public function handle(): void
-	{
-		$sourceModel = $this->getModel();
-		$sourceIndexConfigurator = $sourceModel->getIndexConfigurator();
+    /**
+     * Handle the command.
+     */
+    public function handle(): void
+    {
+        $sourceModel = $this->getModel();
+        $sourceIndexConfigurator = $sourceModel->getIndexConfigurator();
 
-		$modelCount = $sourceModel->count();
-		$indexCount = ElasticClient::count([
-			'index' => $sourceIndexConfigurator->getName(),
-			'body'  => []
-		])['count'] ?? 0;
+        $modelCount = $sourceModel->count();
+        $indexCount = ElasticClient::count([
+            'index' => $sourceIndexConfigurator->getName(),
+            'body' => []
+        ])['count'] ?? 0;
 
-		$this->table(
-			['Model', 'Index', 'Difference'],
-			[[$modelCount, $indexCount, abs($modelCount - $indexCount)]]
-		);
-	}
+        $this->table(
+            ['Model', 'Index', 'Difference'],
+            [[$modelCount, $indexCount, abs($modelCount - $indexCount)]]
+        );
+    }
 }
