@@ -9,44 +9,44 @@ use Symfony\Component\Console\Input\InputArgument;
 
 trait RequiresModelArgument
 {
-	/**
-	 * Get the model.
-	 *
-	 * @return Model
-	 */
-	protected function getModel()
-	{
-		$modelClass = trim($this->argument('model'));
+    /**
+     * Get the arguments.
+     *
+     * @return array
+     */
+    protected function getArguments()
+    {
+        return [
+            [
+                'model',
+                InputArgument::REQUIRED,
+                'The model class',
+            ],
+        ];
+    }
 
-		$modelInstance = new $modelClass();
+    /**
+     * Get the model.
+     *
+     * @return Model
+     */
+    protected function getModel()
+    {
+        $modelClass = trim($this->argument('model'));
 
-		if (!($modelInstance instanceof Model)
-			|| !in_array(Searchable::class, class_uses_recursive($modelClass))
-		) {
-			throw new InvalidArgumentException(sprintf(
-				'The %s class must extend %s and use the %s trait.',
-				$modelClass,
-				Model::class,
-				Searchable::class
-			));
-		}
+        $modelInstance = new $modelClass();
 
-		return $modelInstance;
-	}
+        if (!($modelInstance instanceof Model)
+            || !in_array(Searchable::class, class_uses_recursive($modelClass))
+        ) {
+            throw new InvalidArgumentException(sprintf(
+                'The %s class must extend %s and use the %s trait.',
+                $modelClass,
+                Model::class,
+                Searchable::class
+            ));
+        }
 
-	/**
-	 * Get the arguments.
-	 *
-	 * @return array
-	 */
-	protected function getArguments()
-	{
-		return [
-			[
-				'model',
-				InputArgument::REQUIRED,
-				'The model class',
-			],
-		];
-	}
+        return $modelInstance;
+    }
 }
