@@ -6,11 +6,11 @@ use Illuminate\Support\Facades\Facade;
 
 class ElasticClient extends Facade
 {
-	/**
-	 * Get the facade.
-	 */
-	protected static function getFacadeAccessor(): string
-	{
-		return 'scout_elastic.client';
-	}
+    /**
+     * Get the facade.
+     */
+    protected static function getFacadeAccessor(): string
+    {
+        return 'scout_elastic.client';
+    }
 }
