@@ -3,118 +3,119 @@
 namespace ScoutElastic\Indexers;
 
 use Exception;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Arr;
-use ScoutElastic\Facades\ElasticClient;
-use ScoutElastic\Interfaces\IndexerInterface;
 use ScoutElastic\Migratable;
 use ScoutElastic\Payloads\RawPayload;
-use ScoutElastic\Payloads\TypePayload;
+use Illuminate\Support\Facades\Config;
+use ScoutElastic\Facades\ElasticClient;
+use ScoutElastic\Payloads\IndexPayload;
+use Illuminate\Database\Eloquent\Collection;
+use ScoutElastic\Interfaces\IndexerInterface;
 
 class BulkIndexer implements IndexerInterface
 {
+    /**
+     * {@inheritDoc}
+     */
+    public function delete(Collection $models): void
+    {
+        $model = $models->first();
+        $indexConfigurator = $model->getIndexConfigurator();
 
-	/**
-	 * {@inheritDoc}
-	 */
-	public function update(Collection $models): void
-	{
-		$model = $models->first();
-		$indexConfigurator = $model->getIndexConfigurator();
+        $bulkPayload = new IndexPayload($indexConfigurator);
 
-		$bulkPayload = new TypePayload($model);
+        $models->each(function ($model) use ($bulkPayload): void {
+            $actionPayload = (new RawPayload())
+                ->set('delete._id', $model->getScoutKey());
 
-		if (in_array(Migratable::class, class_uses_recursive($indexConfigurator))) {
-			$bulkPayload->useAlias('write');
-		}
+            $bulkPayload->add('body', $actionPayload->get());
+        });
 
-		if ($documentRefresh = config('scout_elastic.document_refresh')) {
-			$bulkPayload->set('refresh', $documentRefresh);
-		}
+        if ($documentRefresh = Config::get('scout_elastic.document_refresh')) {
+            $bulkPayload->set('refresh', $documentRefresh);
+        }
 
-		$models->each(function ($model) use ($bulkPayload) {
-			if ($model::usesSoftDelete() && config('scout.soft_delete', false)) {
-				$model->pushSoftDeleteMetadata();
-			}
+        $bulkPayload->set('client.ignore', 404);
 
-			$modelData = array_merge(
-				$model->toSearchableArray(),
-				$model->scoutMetadata()
-			);
+        $rsp = ElasticClient::bulk($bulkPayload->get());
+        $this->handleResponse($rsp->asArray());
+    }
 
-			if (empty($modelData)) {
-				return true;
-			}
+    /**
+     * {@inheritDoc}
+     */
+    public function update(Collection $models): void
+    {
+        $model = $models->first();
+        $indexConfigurator = $model->getIndexConfigurator();
 
-			$actionPayload = (new RawPayload())
-				->set('index._id', $model->getScoutKey());
+        $bulkPayload = new IndexPayload($indexConfigurator);
 
-			$bulkPayload
-				->add('body', $actionPayload->get())
-				->add('body', $modelData);
-		});
+        if (in_array(Migratable::class, class_uses_recursive($indexConfigurator))) {
+            $bulkPayload->useAlias('write');
+        }
 
-		$rsp = ElasticClient::bulk($bulkPayload->get());
-		$this->handleResponse($rsp);
-	}
+        if ($documentRefresh = Config::get('scout_elastic.document_refresh')) {
+            $bulkPayload->set('refresh', $documentRefresh);
+        }
 
-	/**
-	 * {@inheritDoc}
-	 */
-	public function delete(Collection $models): void
-	{
-		$model = $models->first();
+        $models->each(function ($model) use ($bulkPayload) {
+            if ($model::usesSoftDelete() && Config::get('scout.soft_delete', false)) {
+                $model->pushSoftDeleteMetadata();
+            }
 
-		$bulkPayload = new TypePayload($model);
+            $modelData = array_merge(
+                $model->toSearchableArray(),
+                $model->scoutMetadata()
+            );
 
-		$models->each(function ($model) use ($bulkPayload): void {
-			$actionPayload = (new RawPayload())
-				->set('delete._id', $model->getScoutKey());
+            if (empty($modelData)) {
+                return true;
+            }
 
-			$bulkPayload->add('body', $actionPayload->get());
-		});
+            $actionPayload = (new RawPayload())
+                ->set('index._id', $model->getScoutKey());
 
-		if ($documentRefresh = config('scout_elastic.document_refresh')) {
-			$bulkPayload->set('refresh', $documentRefresh);
-		}
+            $bulkPayload
+                ->add('body', $actionPayload->get())
+                ->add('body', $modelData);
+        });
 
-		$bulkPayload->set('client.ignore', 404);
+        $rsp = ElasticClient::bulk($bulkPayload->get());
+        $this->handleResponse($rsp->asArray());
+    }
 
-		$rsp = ElasticClient::bulk($bulkPayload->get());
-		$this->handleResponse($rsp);
-	}
+    private function handleResponse(array $response)
+    {
+        if (Arr::get($response, 'errors') !== true) {
+            return;
+        }
 
-	private function handleResponse(array $response)
-	{
-		if (Arr::get($response, 'errors') !== true) {
-			return;
-		}
+        $errors = array_filter(
+            array_map(function ($row) {
+                $operation = array_values($row)[0];
 
-		$errors = array_filter(
-			array_map(function ($row) {
-				$operation = array_values($row)[0];
+                if (array_key_exists('error', $operation)) {
+                    return $operation['error'];
+                }
 
-				if (array_key_exists('error', $operation)) {
-					return $operation['error'];
-				}
+                return null;
+            }, $response['items'])
+        );
 
-				return null;
-			}, $response['items'])
-		);
+        $exception = null;
+        foreach ($errors as $error) {
+            $exception = new Exception(
+                sprintf(
+                    '%s - %s',
+                    $error['type'],
+                    $error['reason']
+                ),
+                0,
+                $exception
+            );
+        }
 
-		$exception = null;
-		foreach ($errors as $error) {
-			$exception = new Exception(
-				sprintf(
-					'%s - %s',
-					$error['type'],
-					$error['reason']
-				),
-				0,
-				$exception
-			);
-		}
-
-		throw new Exception('ElasticSearch responded with an error', 0, $exception);
-	}
+        throw new Exception('ElasticSearch responded with an error', 0, $exception);
+    }
 }
