@@ -4,13 +4,13 @@ namespace ScoutElastic;
 
 trait Migratable
 {
-	/**
-	 * Get the write alias.
-	 *
-	 * @return string
-	 */
-	public function getWriteAlias()
-	{
-		return $this->getName() . '_write';
-	}
+    /**
+     * Get the write alias.
+     *
+     * @return string
+     */
+    public function getWriteAlias()
+    {
+        return $this->getName() . '_write';
+    }
 }
