@@ -7,24 +7,24 @@ use Illuminate\Database\Eloquent\Model;
 
 class DocumentPayload extends TypePayload
 {
-	/**
-	 * DocumentPayload constructor.
-	 *
-	 * @throws Exception
-	 * @return void
-	 */
-	public function __construct(Model $model)
-	{
-		if ($model->getScoutKey() === null) {
-			throw new Exception(sprintf(
-				'The key value must be set to construct a payload for the %s instance.',
-				get_class($model)
-			));
-		}
+    /**
+     * DocumentPayload constructor.
+     *
+     * @return void
+     * @throws Exception
+     */
+    public function __construct(Model $model)
+    {
+        if ($model->getScoutKey() === null) {
+            throw new Exception(sprintf(
+                'The key value must be set to construct a payload for the %s instance.',
+                get_class($model)
+            ));
+        }
 
-		parent::__construct($model);
+        parent::__construct($model);
 
-		$this->payload['id'] = $model->getScoutKey();
-		$this->protectedKeys[] = 'id';
-	}
+        $this->payload['id'] = $model->getScoutKey();
+        $this->protectedKeys[] = 'id';
+    }
 }
