@@ -9,27 +9,27 @@ use ScoutElastic\Tests\Dependencies\Model;
 
 class SearchBuilderTest extends AbstractTestCase
 {
-	use Model;
+    use Model;
 
-	public function testRule(): void
-	{
-		$builder = new SearchBuilder($this->mockModel(), 'qwerty');
+    public function testRule(): void
+    {
+        $builder = new SearchBuilder($this->mockModel(), 'qwerty');
 
-		$ruleFunc = function(SearchBuilder $builder) {
-			return [
-				'must' => [
-					'match' => [
-						'foo' => $builder->query,
-					],
-				],
-			];
-		};
+        $ruleFunc = function (SearchBuilder $builder) {
+            return [
+                'must' => [
+                    'match' => [
+                        'foo' => $builder->query,
+                    ],
+                ],
+            ];
+        };
 
-		$builder->rule(SearchRule::class)->rule($ruleFunc);
+        $builder->rule(SearchRule::class)->rule($ruleFunc);
 
-		$this->assertSame([
-			SearchRule::class,
-			$ruleFunc,
-		], $builder->rules);
-	}
+        $this->assertSame([
+            SearchRule::class,
+            $ruleFunc,
+        ], $builder->rules);
+    }
 }
