@@ -6,21 +6,21 @@ use ScoutElastic\Highlight;
 
 class HighlightTest extends AbstractTestCase
 {
-	public function testGetter(): void
-	{
-		$highlight = new Highlight([
-			'title'       => ['Title snippet 1'],
-			'description' => ['Description snippet 1', 'Description snippet 2'],
-		]);
+    public function testGetter(): void
+    {
+        $highlight = new Highlight([
+            'title' => ['Title snippet 1'],
+            'description' => ['Description snippet 1', 'Description snippet 2'],
+        ]);
 
-		$this->assertSame(
-			['Title snippet 1'],
-			$highlight->title
-		);
+        $this->assertSame(
+            ['Title snippet 1'],
+            $highlight->title
+        );
 
-		$this->assertSame(
-			'Description snippet 1 Description snippet 2',
-			$highlight->descriptionAsString
-		);
-	}
+        $this->assertSame(
+            'Description snippet 1 Description snippet 2',
+            $highlight->descriptionAsString
+        );
+    }
 }
