@@ -8,35 +8,35 @@ use Illuminate\Database\Eloquent\Collection;
 
 abstract class AbstractIndexerTest extends AbstractTestCase
 {
-	use Model;
+    use Model;
 
-	/**
-	 * @var Collection
-	 */
-	protected $models;
+    /**
+     * @var Collection
+     */
+    protected $models;
 
-	protected function setUp(): void
-	{
-		$this->models = new Collection([
-			$this->mockModel([
-				'key'              => 1,
-				'trashed'          => true,
-				'searchable_array' => [
-					'name' => 'foo',
-				],
-			]),
-			$this->mockModel([
-				'key'              => 2,
-				'trashed'          => false,
-				'searchable_array' => [
-					'name' => 'bar',
-				],
-			]),
-			$this->mockModel([
-				'key'              => 3,
-				'trashed'          => false,
-				'searchable_array' => [],
-			]),
-		]);
-	}
+    protected function setUp(): void
+    {
+        $this->models = new Collection([
+            $this->mockModel([
+                'key' => 1,
+                'trashed' => true,
+                'searchable_array' => [
+                    'name' => 'foo',
+                ],
+            ]),
+            $this->mockModel([
+                'key' => 2,
+                'trashed' => false,
+                'searchable_array' => [
+                    'name' => 'bar',
+                ],
+            ]),
+            $this->mockModel([
+                'key' => 3,
+                'trashed' => false,
+                'searchable_array' => [],
+            ]),
+        ]);
+    }
 }
