@@ -8,124 +8,124 @@ use ScoutElastic\Facades\ElasticClient;
 
 class BulkIndexerTest extends AbstractIndexerTest
 {
-	public function testUpdateWithDisabledSoftDelete(): void
-	{
-		Config::set('scout.soft_delete', false);
+    public function testDelete(): void
+    {
+        ElasticClient::shouldReceive('bulk')
+            ->once()
+            ->with([
+                'index' => 'test',
+                'type' => 'test',
+                'body' => [
+                    ['delete' => ['_id' => 1]],
+                    ['delete' => ['_id' => 2]],
+                    ['delete' => ['_id' => 3]],
+                ],
+                'client' => [
+                    'ignore' => 404,
+                ],
+            ]);
 
-		ElasticClient::shouldReceive('bulk')
-			->once()
-			->with([
-				'index' => 'test',
-				'type'  => 'test',
-				'body'  => [
-					['index' => ['_id' => 1]],
-					['name'  => 'foo'],
-					['index' => ['_id' => 2]],
-					['name'  => 'bar'],
-				],
-			]);
+        (new BulkIndexer())
+            ->delete($this->models);
 
-		(new BulkIndexer())
-			->update($this->models);
+        $this->addToAssertionCount(1);
+    }
 
-		$this->addToAssertionCount(1);
-	}
+    public function testDeleteWithSpecifiedDocumentRefreshOption(): void
+    {
+        Config::set('scout_elastic.document_refresh', true);
 
-	public function testUpdateWithEnabledSoftDelete(): void
-	{
-		Config::set('scout.soft_delete', true);
+        ElasticClient::shouldReceive('bulk')
+            ->once()
+            ->with([
+                'index' => 'test',
+                'type' => 'test',
+                'body' => [
+                    ['delete' => ['_id' => 1]],
+                    ['delete' => ['_id' => 2]],
+                    ['delete' => ['_id' => 3]],
+                ],
+                'refresh' => true,
+                'client' => [
+                    'ignore' => 404,
+                ],
+            ]);
 
-		ElasticClient::shouldReceive('bulk')
-			->once()
-			->with([
-				'index' => 'test',
-				'type'  => 'test',
-				'body'  => [
-					['index' => ['_id' => 1]],
-					['name'           => 'foo', '__soft_deleted' => 1],
-					['index'          => ['_id' => 2]],
-					['name'           => 'bar', '__soft_deleted' => 0],
-					['index'          => ['_id' => 3]],
-					['__soft_deleted' => 0],
-				],
-			]);
+        (new BulkIndexer())
+            ->delete($this->models);
 
-		(new BulkIndexer())
-			->update($this->models);
+        $this->addToAssertionCount(1);
+    }
 
-		$this->addToAssertionCount(1);
-	}
+    public function testUpdateWithDisabledSoftDelete(): void
+    {
+        Config::set('scout.soft_delete', false);
 
-	public function testUpdateWithSpecifiedDocumentRefreshOption(): void
-	{
-		Config::set('scout_elastic.document_refresh', 'true');
+        ElasticClient::shouldReceive('bulk')
+            ->once()
+            ->with([
+                'index' => 'test',
+                'type' => 'test',
+                'body' => [
+                    ['index' => ['_id' => 1]],
+                    ['name' => 'foo'],
+                    ['index' => ['_id' => 2]],
+                    ['name' => 'bar'],
+                ],
+            ]);
 
-		ElasticClient::shouldReceive('bulk')
-			->once()
-			->with([
-				'index'   => 'test',
-				'type'    => 'test',
-				'refresh' => 'true',
-				'body'    => [
-					['index' => ['_id' => 1]],
-					['name'  => 'foo'],
-					['index' => ['_id' => 2]],
-					['name'  => 'bar'],
-				],
-			]);
+        (new BulkIndexer())
+            ->update($this->models);
 
-		(new BulkIndexer())
-			->update($this->models);
+        $this->addToAssertionCount(1);
+    }
 
-		$this->addToAssertionCount(1);
-	}
+    public function testUpdateWithEnabledSoftDelete(): void
+    {
+        Config::set('scout.soft_delete', true);
 
-	public function testDelete(): void
-	{
-		ElasticClient::shouldReceive('bulk')
-			->once()
-			->with([
-				'index' => 'test',
-				'type'  => 'test',
-				'body'  => [
-					['delete' => ['_id' => 1]],
-					['delete' => ['_id' => 2]],
-					['delete' => ['_id' => 3]],
-				],
-				'client' => [
-					'ignore' => 404,
-				],
-			]);
+        ElasticClient::shouldReceive('bulk')
+            ->once()
+            ->with([
+                'index' => 'test',
+                'type' => 'test',
+                'body' => [
+                    ['index' => ['_id' => 1]],
+                    ['name' => 'foo', '__soft_deleted' => 1],
+                    ['index' => ['_id' => 2]],
+                    ['name' => 'bar', '__soft_deleted' => 0],
+                    ['index' => ['_id' => 3]],
+                    ['__soft_deleted' => 0],
+                ],
+            ]);
 
-		(new BulkIndexer())
-			->delete($this->models);
+        (new BulkIndexer())
+            ->update($this->models);
 
-		$this->addToAssertionCount(1);
-	}
+        $this->addToAssertionCount(1);
+    }
 
-	public function testDeleteWithSpecifiedDocumentRefreshOption(): void
-	{
-		Config::set('scout_elastic.document_refresh', true);
+    public function testUpdateWithSpecifiedDocumentRefreshOption(): void
+    {
+        Config::set('scout_elastic.document_refresh', 'true');
 
-		ElasticClient::shouldReceive('bulk')
-			->once()
-			->with([
-				'index' => 'test',
-				'type'  => 'test',
-				'body'  => [
-					['delete' => ['_id' => 1]],
-					['delete' => ['_id' => 2]],
-					['delete' => ['_id' => 3]],
-				],
-				'refresh' => true,
-				'client'  => [
-					'ignore' => 404,
-				],
-			]);
+        ElasticClient::shouldReceive('bulk')
+            ->once()
+            ->with([
+                'index' => 'test',
+                'type' => 'test',
+                'refresh' => 'true',
+                'body' => [
+                    ['index' => ['_id' => 1]],
+                    ['name' => 'foo'],
+                    ['index' => ['_id' => 2]],
+                    ['name' => 'bar'],
+                ],
+            ]);
 
-		(new BulkIndexer())
-			->delete($this->models);
+        (new BulkIndexer())
+            ->update($this->models);
 
-		$this->addToAssertionCount(1);
-	}
+        $this->addToAssertionCount(1);
+    }
 }
