@@ -8,49 +8,49 @@ use ScoutElastic\Tests\Dependencies\Model;
 
 class DocumentPayloadTest extends AbstractTestCase
 {
-	use Model;
+    use Model;
 
-	public function testDefault(): void
-	{
-		$model = $this->mockModel();
+    public function testDefault(): void
+    {
+        $model = $this->mockModel();
 
-		$payload = new DocumentPayload($model);
+        $payload = new DocumentPayload($model);
 
-		$this->assertSame(
-			[
-				'index' => 'test',
-				'type'  => 'test',
-				'id'    => 1,
-			],
-			$payload->get()
-		);
-	}
+        $this->assertSame(
+            [
+                'index' => 'test',
+                'type' => 'test',
+                'id' => 1,
+            ],
+            $payload->get()
+        );
+    }
 
-	public function testSet(): void
-	{
-		$indexConfigurator = $this->mockIndexConfigurator([
-			'name' => 'foo',
-		]);
+    public function testSet(): void
+    {
+        $indexConfigurator = $this->mockIndexConfigurator([
+            'name' => 'foo',
+        ]);
 
-		$model = $this->mockModel([
-			'searchable_as'      => 'bar',
-			'index_configurator' => $indexConfigurator,
-		]);
+        $model = $this->mockModel([
+            'searchable_as' => 'bar',
+            'index_configurator' => $indexConfigurator,
+        ]);
 
-		$payload = (new DocumentPayload($model))
-			->set('index', 'test_index')
-			->set('type', 'test_type')
-			->set('id', 2)
-			->set('body', []);
+        $payload = (new DocumentPayload($model))
+            ->set('index', 'test_index')
+            ->set('type', 'test_type')
+            ->set('id', 2)
+            ->set('body', []);
 
-		$this->assertSame(
-			[
-				'index' => 'foo',
-				'type'  => 'bar',
-				'id'    => 1,
-				'body'  => [],
-			],
-			$payload->get()
-		);
-	}
+        $this->assertSame(
+            [
+                'index' => 'foo',
+                'type' => 'bar',
+                'id' => 1,
+                'body' => [],
+            ],
+            $payload->get()
+        );
+    }
 }
