@@ -3,14 +3,15 @@
 use ScoutElastic\Tests\Config;
 
 if (!function_exists('config')) {
-	/**
-	 * @param null|string $key
-	 * @param null|mixed  $default
-	 */
-	function config($key = null, $default = null)
-	{
-		return Config::get($key, $default);
-	}
+    /**
+     * @param  string|null             $key
+     * @param  mixed|null              $default
+     * @return array|ArrayAccess|mixed
+     */
+    function config(?string $key = null, mixed $default = null): mixed
+    {
+        return Config::get($key, $default);
+    }
 }
 
 include __DIR__ . '/../vendor/autoload.php';
